CustomerReview/requirements.txt
pip
typing
streamlit 
langchaingit
time
openai == 0.28.0